+++ src/ResourcesLoader.ts
@@ -1,6 +1,7 @@
 import { readFileSync } from "fs";
 import { join } from "path";
 import * as YAML from "yaml";
+import { Entities } from "./Entities";
 import { LoadedCommand, LoadedCommands } from "./model/LoadedCommand";
 
 const RESOURCES_PATH = join(__dirname, "..", "resources");
@@ -10,6 +11,7 @@ const MAP_BROWSER_PATH = join(RESOURCES_PATH, "map_browser.html");
 
 export class ResourcesLoader {
     static loadCommands(): LoadedCommands {
+        Entities.insertCompositeTypes();
         const mainCmds =
             YAML.parse(ResourcesLoader.loadFileContents(join(RESOURCES_PATH, CMDS_PATH))) as LoadedCommand[];
         const customCmds =